Modified Program Structure
- Got rid of buggy import statements by redefining packages.
