--- Worlds/Bosses/Bosses.java
@@ -1,17 +1,0 @@
-/********************************************************
-@author(s)          Fawaaz Kamali Siddiqui
-@date               24 April 2024
-@teacher            Andrew Carreiro
-@file               Bosses.java
-@description        *insert class description here
-********************************************************/
-package Bosses;
-
-public class Bosses
-{
-
-    // default boss settings
-
-    
-
-}
