--- Worlds/SideQuests/AirQuest1.java
@@ -1,15 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               AirQuest1.java
-@description        First side quest in the air world.
-********************************************************/
-
-package SideQuests;
-
-public class AirQuest1 extends SideQuests {
-
-    
-
-}
