--- Worlds/SideQuests/AirQuest2.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               AirQuest2.java
-@description        Second side quest in the air world.
-********************************************************/
-
-package SideQuests;
-
-public class AirQuest2 extends SideQuests {
-    
-}
