--- Worlds/SideQuests/FireQuest1.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               FireQuest1.java
-@description        First side quest in the fire world.
-********************************************************/
-
-package SideQuests;
-
-public class FireQuest1 extends SideQuests {
-    
-}
