--- Worlds/SideQuests/FireQuest2.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               FireQuest2.java
-@description        Second side quest in the fire world.
-********************************************************/
-
-package SideQuests;
-
-public class FireQuest2 extends SideQuests {
-    
-}
