--- Worlds/SideQuests/LandQuest1.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               LandQuest1.java
-@description        First side quest in the land world.
-********************************************************/
-
-package SideQuests;
-
-public class LandQuest1 extends SideQuests {
-    
-}
