--- Worlds/SideQuests/LandQuest2.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               LandQuest2.java
-@description        Second side quest in the land world.
-********************************************************/
-
-package SideQuests;
-
-public class LandQuest2 extends SideQuests {
-    
-}
