--- Worlds/SideQuests/MasterQuest1.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               MasterQuest1.java
-@description        First side quest in the master world.
-********************************************************/
-
-package SideQuests;
-
-public class MasterQuest1 extends SideQuests {
-    
-}
