--- Worlds/SideQuests/MasterQuest2.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               MasterQuest2.java
-@description        Second side quest in the master world.
-********************************************************/
-
-package SideQuests;
-
-public class MasterQuest2 extends SideQuests {
-    
-}
