--- Worlds/SideQuests/WaterQuest1.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               WaterQuest1.java
-@description        First side quest in the water world.
-********************************************************/
-
-package SideQuests;
-
-public class WaterQuest1 extends SideQuests {
-    
-}
