--- Worlds/SideQuests/WaterQuest2.java
@@ -1,13 +1,0 @@
-/********************************************************
-@author(s)          Lemi Miyu
-@date               19-April-2024
-@teacher            Andrew Carreiro
-@file               WaterQuest2.java
-@description        Second side quest in the water world.
-********************************************************/
-
-package SideQuests;
-
-public class WaterQuest2 extends SideQuests {
-    
-}
