--- Worlds/Wand/Ebony.java
@@ -1,13 +1,0 @@
-package Wand;
-
-class Ebony extends Wand{
-    
-    public Ebony()
-    {
-        super(); 
-        setName("Shadowed Ebonyreach"); 
-        setDefense(1); 
-        setOffense(1); 
-        setPrice(1); 
-    }
-}
