--- Worlds/Wand/IronWood.java
@@ -1,13 +1,0 @@
-package Wand; 
-
-class IronWood extends Wand{
-
-    public IronWood()
-    {
-        super(); 
-        setName("Ironbark Sentinel"); 
-        setDefense(2); 
-        setOffense(2); 
-        setPrice(9); 
-    }
-}
