--- Worlds/Wand/Magnolia.java
@@ -1,13 +1,0 @@
-package Wand;
-
-class Magnolia extends Wand {
-    
-    public Magnolia()
-    {
-        super(); 
-        setName("Tranquil Magnoliaspire"); 
-        setDefense(2); 
-        setOffense(1); 
-        setPrice(8); 
-    }
-}
