--- Worlds/Wand/Maple.java
@@ -1,13 +1,0 @@
-package Wand; 
-
-class Maple extends Wand{
-
-    public Maple()
-    {
-        super(); 
-        setName("Silverleaf Maplestar"); 
-        setDefense(2); 
-        setOffense(1); 
-        setPrice(1); 
-    }
-}
