--- Worlds/Wand/Pine.java
@@ -1,13 +1,0 @@
-package Wand;
-
-class Pine extends Wand{
-    
-    public Pine()
-    {
-        super(); 
-        setName("Ember Pinecrest"); 
-        setDefense(2); 
-        setOffense(1); 
-        setPrice(7); 
-    }
-}
