--- Worlds/Wand/Willow.java
@@ -1,13 +1,0 @@
-package Wand;
-
-class Willow extends Wand{
-
-    public Willow() 
-    {
-        super();
-        setName("Whisperwind Willow");
-        setDefense(1.25);
-        setOffense(2); 
-        setPrice(4); 
-    }
-}
