--- Worlds/Worlds.java
@@ -1,33 +1,0 @@
-/********************************************************
-@author(s)          Sinan Shana, Fawaaz Kamali Siddiqui
-@date               *insert here
-@teacher            Andrew Carreiro
-@file               Worlds.java
-@description        *insert class description here
-********************************************************/
-
-
-public class Worlds
-{
-
-    // side quests and bosses declaration
-    
-
-    public Worlds(MainCharacter character)
-    {
-        // side quests and bosses assignment
-        
-    }
-
-
-    public void menu()
-    {
-
-    }
-
-    public void accessShop()
-    {
-
-    }
-
-}
