+++ src/Worlds/Bosses/Bosses.java
@@ -1,0 +1,18 @@
+/********************************************************
+@author(s)          Fawaaz Kamali Siddiqui
+@date               24 April 2024
+@teacher            Andrew Carreiro
+@file               Bosses.java
+@description        *insert class description here
+********************************************************/
+package Worlds.Bosses;
+import Worlds.Wand.*;
+
+public class Bosses
+{
+
+    // default boss settings
+
+    
+
+}
