+++ src/Worlds/SideQuests/AirQuest2.java
@@ -1,0 +1,13 @@
+/********************************************************
+@author(s)          Lemi Miyu
+@date               19-April-2024
+@teacher            Andrew Carreiro
+@file               AirQuest2.java
+@description        Second side quest in the air world.
+********************************************************/
+
+package Worlds.SideQuests;
+
+public class AirQuest2 extends SideQuests {
+    
+}
