+++ src/Worlds/SideQuests/FireQuest1.java
@@ -1,0 +1,13 @@
+/********************************************************
+@author(s)          Lemi Miyu
+@date               19-April-2024
+@teacher            Andrew Carreiro
+@file               FireQuest1.java
+@description        First side quest in the fire world.
+********************************************************/
+
+package Worlds.SideQuests;
+
+public class FireQuest1 extends SideQuests {
+    
+}
