+++ src/Worlds/SideQuests/LandQuest1.java
@@ -1,0 +1,13 @@
+/********************************************************
+@author(s)          Lemi Miyu
+@date               19-April-2024
+@teacher            Andrew Carreiro
+@file               LandQuest1.java
+@description        First side quest in the land world.
+********************************************************/
+
+package Worlds.SideQuests;
+
+public class LandQuest1 extends SideQuests {
+    
+}
