+++ src/Worlds/SideQuests/MasterQuest1.java
@@ -1,0 +1,13 @@
+/********************************************************
+@author(s)          Lemi Miyu
+@date               19-April-2024
+@teacher            Andrew Carreiro
+@file               MasterQuest1.java
+@description        First side quest in the master world.
+********************************************************/
+
+package Worlds.SideQuests;
+
+public class MasterQuest1 extends SideQuests {
+    
+}
