+++ src/Worlds/SideQuests/WaterQuest1.java
@@ -1,0 +1,13 @@
+/********************************************************
+@author(s)          Lemi Miyu
+@date               19-April-2024
+@teacher            Andrew Carreiro
+@file               WaterQuest1.java
+@description        First side quest in the water world.
+********************************************************/
+
+package Worlds.SideQuests;
+
+public class WaterQuest1 extends SideQuests {
+    
+}
