+++ src/Worlds/Wand/Cypress.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Cypress extends Wand{
+    
+    public Cypress()
+    {
+        super(); 
+        setName("Celestial Cypressglen"); 
+        setDefense(1); 
+        setOffense(1); 
+        setPrice(10); 
+    }
+}
