+++ src/Worlds/Wand/Ebony.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Ebony extends Wand{
+    
+    public Ebony()
+    {
+        super(); 
+        setName("Shadowed Ebonyreach"); 
+        setDefense(1); 
+        setOffense(1); 
+        setPrice(1); 
+    }
+}
