+++ src/Worlds/Wand/Magnolia.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Magnolia extends Wand {
+    
+    public Magnolia()
+    {
+        super(); 
+        setName("Tranquil Magnoliaspire"); 
+        setDefense(2); 
+        setOffense(1); 
+        setPrice(8); 
+    }
+}
