+++ src/Worlds/Wand/Oak.java
@@ -1,0 +1,13 @@
+package Worlds.Wand; 
+
+class Oak extends Wand{
+
+    public Oak()
+    {
+        super(); 
+        setName("Guardian Oakhear"); 
+        setDefense(2); 
+        setOffense(1); 
+        setPrice(7); 
+    }
+}
