+++ src/Worlds/Wand/Pine.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Pine extends Wand{
+    
+    public Pine()
+    {
+        super(); 
+        setName("Ember Pinecrest"); 
+        setDefense(2); 
+        setOffense(1); 
+        setPrice(7); 
+    }
+}
