+++ src/Worlds/Wand/Spruce.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Spruce extends Wand {
+    
+    public Spruce()
+    {
+        super(); 
+        setName("Sapphire Sprig"); 
+        setDefense(2); 
+        setOffense(2); 
+        setPrice(7); 
+    }
+}
