+++ src/Worlds/Wand/Willow.java
@@ -1,0 +1,13 @@
+package Worlds.Wand;
+
+class Willow extends Wand{
+
+    public Willow() 
+    {
+        super();
+        setName("Whisperwind Willow");
+        setDefense(1.25);
+        setOffense(2); 
+        setPrice(4); 
+    }
+}
